Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/drawing-rectangles-to-the-canvas/2d.clearRect.clip-expected.txt

2d.clearRect.clip
clearRect is affected by clipping regions
Actual output:
Expected output:


PASS clearRect is affected by clipping regions 
